Use the WAF as dynamic library
Remove the code to merge the .lib / C++ coee into the
ClrProfiler.Native.[dll|so|dynlib].

This means we now need to be able to find the plaform specific library,
so I've added some code for this.

It also changes the way the rule file is stored. Resource is probably
better than a massive string constant.